requirements.txt
tensorflow==2.13.0
pillow>=9.0.0
gradio==3.50.2
numpy>=1.24.0
